backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
aiosmtpd==1.4.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
aiosmtplib==3.0.2
bleach==6.1.0
bcrypt==4.0.1
redis==5.1.1
celery[redis]==5.4.0
dnspython>=2.6.0
dkimpy==1.1.8
pyspf==2.0.14
email-validator==2.2.0
python-multipart==0.0.12
orjson==3.10.7
structlog==24.4.0
httpx>=0.27.0
